+++ frontend/src/components/plugins/tables/pluginTableColumns.jsx
@@ -21,12 +21,12 @@ import {
   PluginDeletionButton,
   PluginConfigButton,
   PlaybookFlowsButton,
+  MappingDataModel,
 } from "./pluginActionsButtons";
 import { JobTypes } from "../../../constants/jobConst";
 import TableCell from "../../common/TableCell";
 import TableCellList from "../../common/TableCellList";
 import { TableCellCollapse } from "../../common/TableCellCollapse";
-import { MappingDataModel } from "./utils";
 
 /* This function is available in the certego-ui, but it doesn't works:
 Only the filter value "all" is shown the filter values that should be generated by the data are not generated.
